disable encoding from slice
