+++ filter/src/protobuf/decode.rs
@@ -1,0 +1,3 @@
+pub struct SubscribeUpdateLimited {
+    //
+}
